PR fixes with updated tests
